move bearer token code
