+++ src/eduid/maccapi/config.py
@@ -1,28 +1,23 @@
 import logging
 from pathlib import Path
+from typing import Optional
 
 from pydantic import Field, validator
 
-from eduid.common.config.base import LoggingConfigMixin, RootConfig
+from eduid.common.config.base import AuthnBearerTokenConfig, LoggingConfigMixin
 from eduid.common.utils import removesuffix
 
 logger = logging.getLogger(__name__)
 
 
-class MAccApiConfig(RootConfig, LoggingConfigMixin):
+class MAccApiConfig(AuthnBearerTokenConfig, LoggingConfigMixin):
     """
     Configuration for the Managed Accounts API app
     """
 
-    protocol: str = "http"
-    server_name: str = "localhost:8000"
-    application_root: str = ""
-    log_format: str = "{asctime} | {levelname:7} | {hostname} | {name:35} | {module:10} | {message}"
-    no_authn_urls: list[str] = Field(default=["^/status/healthy$", "^/docs/?$", "^/openapi.json"])
     vccs_url: str = "http://vccs:8080/"
-    mongo_uri: str = ""
-    authorization_mandatory: bool = True
-    keystore_path: Path
+    # The expected value of the authn JWT claims['requested_access']['type']
+    requested_access_type: Optional[str] = "maccapi"
 
     @validator("application_root")
     def application_root_must_not_end_with_slash(cls, v: str):
