+++ src/eduid/scimapi/routers/login.py
@@ -3,10 +3,10 @@ import datetime
 from fastapi import Response
 from jwcrypto import jwt
 
+from eduid.common.models.webauthn import AuthSource
 from eduid.scimapi.api_router import APIRouter
 from eduid.scimapi.context_request import ContextRequest
 from eduid.scimapi.exceptions import ErrorDetail, NotFound, Unauthorized
-from eduid.scimapi.middleware import AuthSource
 from eduid.scimapi.models.login import TokenRequest
 
 login_router = APIRouter(
