+++ src/eduid/scimapi/routers/utils/events.py
@@ -4,9 +4,9 @@ from uuid import uuid4
 
 from fastapi import Response
 
+from eduid.common.config.base import DataOwnerName
 from eduid.common.models.scim_base import Meta, SCIMResourceType, SCIMSchema, WeakVersion
 from eduid.common.utils import urlappend
-from eduid.scimapi.config import DataOwnerName
 from eduid.scimapi.context_request import ContextRequest
 from eduid.scimapi.exceptions import BadRequest
 from eduid.scimapi.models.event import EventResponse, NutidEventExtensionV1, NutidEventResource
